+++ .gitignore
@@ -115,3 +115,6 @@ Thumbs.db
 Python27
 python
 Python*
+python-2.7.1-32
+python-2.7.1-64
+python-3.3.1-64
