+++ README.md
@@ -6,7 +6,7 @@
  
  	Author: Subin. Gopi (subing85@gmail.com)
  	
- 	version 0.0.2 linux
+ 	version 0.0.2 linux
