+++ studio/launcherPTD/main.sh
@@ -1,0 +1,1 @@
+print 'sssssssssssssss'
