+++ studio/tmp/launcher.bat
