+++ studio/tmp/launcherPTD.bat
